Errors and requests were moved to new files.
